requirements_stats_aware_stylegan2-ada.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pyhd8ed1ab_0
blas=1.0=mkl
bottleneck=1.3.5=py37h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py37h27cfd23_1003
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.12.12=h06a4308_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=py37h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py37h06a4308_0
colorama=0.4.6=py37h06a4308_0
cryptography=39.0.1=py37h9ce1e76_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.11.0=pyhd3eb1b0_0
fftw=3.3.9=h27cfd23_1
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
grpcio=1.42.0=py37hce63b2e_0
idna=3.4=py37h06a4308_0
importlib-metadata=4.11.3=py37h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.1=py37h06a4308_0
jpeg=9b=h024ee3a_2
kiwisolver=1.4.4=py37h6a678d5_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.39=h5eee18b_0
libprotobuf=3.15.8=h780b84a_1
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=h85742a9_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.4=pyhd8ed1ab_0
matplotlib-base=3.5.3=py37hf590b9c_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
ninja=1.11.1=pypi_0
ninja-base=1.10.2=hd09550d_5
numexpr=2.8.4=py37he184ba9_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
openssl=1.1.1w=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=22.0=py37h06a4308_0
pandas=1.3.5=py37h8c16a72_0
pillow=9.3.0=py37hace64e9_1
pip=22.3.1=py37h06a4308_0
protobuf=3.15.8=py37hcd2ae1e_0
psutil=5.9.0=py37h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py37h06a4308_0
pyparsing=3.0.9=py37h06a4308_0
pyro-api=0.1.2=pyhd8ed1ab_0
pyro-ppl=1.5.2=pyhd8ed1ab_0
pysocks=1.7.1=py37_1
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.7.1=py3.7_cuda11.0.221_cudnn8.0.5_0
pytz=2022.7=py37h06a4308_0
readline=8.2=h5eee18b_0
requests=2.28.1=py37h06a4308_0
scikit-learn=1.0.2=py37h51133e4_1
scipy=1.7.3=py37h6c91a56_2
seaborn=0.12.2=py37h06a4308_0
setuptools=65.6.3=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
tensorboard=1.15.0=py37_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
torch-two-sample=0.1=pypi_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu110
tqdm=4.66.2=pyhd8ed1ab_0
typing_extensions=4.4.0=py37h06a4308_0
urllib3=1.26.14=py37h06a4308_0
werkzeug=2.1.2=pyhd8ed1ab_1
wheel=0.38.4=py37h06a4308_0
xz=5.4.2=h5eee18b_0
zipp=3.11.0=py37h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=haebb681_0
